# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Event Finder API - discover charity galas, silent auctions, and fundraising events
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: List discovered events with filtering
      parameters:
        - name: runId
          in: query
          schema:
            type: ["integer", "null"]
        - name: tier
          in: query
          schema:
            type: ["string", "null"]
        - name: auctionType
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: ["NEW", "UPDATED", "UNCHANGED", null]
        - name: minScore
          in: query
          schema:
            type: ["integer", "null"]
        - name: maxScore
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: sortBy
          in: query
          schema:
            type: ["string", "null"]
        - name: sortDir
          in: query
          schema:
            type: ["string", "null"]
            enum: ["asc", "desc", null]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: hidePast
          in: query
          schema:
            type: ["boolean", "null"]
        - name: pastOnly
          in: query
          description: When true, return only past events (overrides hidePast).
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsPage"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/export:
    get:
      operationId: exportEvents
      tags: [events]
      summary: Export events as CSV
      parameters:
        - name: runId
          in: query
          schema:
            type: ["integer", "null"]
        - name: tier
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

  # ─── Dashboard ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, runId, score, tier, status, createdAt]
      properties:
        id:                   { type: integer }
        runId:                { type: integer }
        score:                { type: integer }
        tier:
          type: string
        status:
          type: string
          enum: [NEW, UPDATED, UNCHANGED]
        eventName:            { type: ["string", "null"] }
        eventDate:            { type: ["string", "null"] }
        eventVenue:           { type: ["string", "null"] }
        eventAddress:         { type: ["string", "null"] }
        eventDescription:     { type: ["string", "null"] }
        auctionType:          { type: ["string", "null"] }
        hasSilentAuction:     { type: ["boolean", "null"] }
        hasLiveAuction:       { type: ["boolean", "null"] }
        ticketPrice:          { type: ["number", "null"] }
        tablePrice:           { type: ["number", "null"] }
        formality:            { type: ["string", "null"] }
        rsvpLink:             { type: ["string", "null"] }
        contactName:          { type: ["string", "null"] }
        contactEmail:         { type: ["string", "null"] }
        contactPhone:         { type: ["string", "null"] }
        eventPageUrl:         { type: ["string", "null"] }
        audienceNote:         { type: ["string", "null"] }
        sponsorshipMentioned: { type: ["boolean", "null"] }
        orgName:              { type: ["string", "null"] }
        orgAddress:           { type: ["string", "null"] }
        orgCity:              { type: ["string", "null"] }
        orgPhone:             { type: ["string", "null"] }
        orgEmail:             { type: ["string", "null"] }
        orgWebsite:           { type: ["string", "null"] }
        isPast:               { type: boolean }
        errorMessage:         { type: ["string", "null"] }
        runName:              { type: ["string", "null"] }
        createdAt:            { type: string }
    EventsPage:
      type: object
      required: [events, total, page, limit]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:  { type: integer }
        page:   { type: integer }
        limit:  { type: integer }

    # ─── Dashboard ────────────────────────────────────────────────────────────────
...